重新启用了tof 打开了按钮的上拉电阻 增加了中断引脚的上拉

+++ Core/Inc/tof.h
@@ -1,4 +1,5 @@
 #pragma once
+#include <stdint.h>
 void TofInit();
 void TofHandler();
-float TofGetDistance();
+uint16_t TofGetDistance();
